+++ pkg/prometheus_ingester/query_executor.go
@@ -4,17 +4,19 @@ import (
 	"context"
 	"fmt"
 	"github.com/hashicorp/go-multierror"
+	"go.uber.org/atomic"
 	"math"
 	"sort"
 	"strconv"
+	"strings"
 	"sync"
 	"time"
 
 	v1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	"github.com/prometheus/common/model"
-	"github.com/sirupsen/logrus"
 	"github.com/seznam/slo-exporter/pkg/event"
 	"github.com/seznam/slo-exporter/pkg/stringmap"
+	"github.com/sirupsen/logrus"
 )
 
 type queryExecutor struct {
@@ -23,8 +25,10 @@ type queryExecutor struct {
 	eventsChan        chan *event.Raw
 	logger            logrus.FieldLogger
 	api               v1.API
+	queryInProgress   atomic.Bool
 	previousResult    queryResult
 	previousResultMtx sync.RWMutex
+	staleness         time.Duration
 }
 
 type queryResult struct {
@@ -32,6 +36,29 @@ type queryResult struct {
 	timestamp time.Time
 	// metric: <most recent sample>
 	metrics map[model.Fingerprint]model.SamplePair
+}
+
+func (r *queryResult) String() string {
+	var res []string
+	for k, v := range r.metrics {
+		res = append(res, fmt.Sprintln(v.Timestamp, k, v.Value))
+	}
+	return strings.Join(res, ",")
+}
+
+func (r *queryResult) dropStaleResults(staleness time.Duration, moment time.Time) {
+	for m, s := range r.metrics {
+		if moment.Sub(s.Timestamp.Time()) > staleness {
+			delete(r.metrics, m)
+		}
+	}
+}
+
+func (r *queryResult) update(new queryResult) {
+	r.timestamp = new.timestamp
+	for m, s := range new.metrics {
+		r.metrics[m] = s
+	}
 }
 
 // withRangeSelector returns q.query concatenated with desired range selector
@@ -48,6 +75,8 @@ func (q *queryExecutor) withRangeSelector(ts time.Time) string {
 
 // execute query at provided timestamp ts
 func (q *queryExecutor) execute(ts time.Time) (model.Value, error) {
+	q.queryInProgress.Store(true)
+	defer q.queryInProgress.Store(false)
 	timeoutCtx, cancel := context.WithTimeout(context.Background(), q.queryTimeout)
 	defer cancel()
 	var (
@@ -90,6 +119,10 @@ func (q *queryExecutor) run(ctx context.Context, wg *sync.WaitGroup) {
 		select {
 		// Wait for the tick
 		case <-ticker.C:
+			if q.queryInProgress.Load() {
+				q.logger.Warn("skipping query execution, previous query still in progress...")
+				continue
+			}
 			ts := time.Now()
 			result, err := q.execute(ts)
 			if err != nil {
@@ -247,6 +280,8 @@ func (q *queryExecutor) processCountersIncrease(matrix model.Matrix, ts time.Tim
 }
 
 func (q *queryExecutor) processMatrixResultAsIncrease(matrix model.Matrix, ts time.Time) chan metricIncrease {
+	// Drop outdated samples from previous result same as prometheus does see https://prometheus.io/docs/prometheus/latest/querying/basics/#staleness
+	q.previousResult.dropStaleResults(q.staleness, ts)
 	outChan := make(chan metricIncrease)
 	go func() {
 		defer close(outChan)
@@ -282,7 +317,7 @@ func (q *queryExecutor) processMatrixResultAsIncrease(matrix model.Matrix, ts ti
 				metric:   singleMetricSampleStream.Metric,
 			}
 		}
-		q.previousResult = currentResult
+		q.previousResult.update(currentResult)
 	}()
 	return outChan
 }