+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [#78](https://github.com/seznam/slo-exporter/pull/78) Upgrade dependencies mainly to avoid reported CVEs in grafana/loki
+- [#80](https://github.com/seznam/slo-exporter/pull/80) prometheus-ingester: Add staleness support and possibility to have shorter query interval
 
 ## [v6.10.0] 2021-09-17
 ### Added